+++ packages/ui/src/stories/molecules/Search/Search.stories.ts
@@ -20,6 +20,7 @@ const meta = {
     isSearchError: { control: 'boolean' },
     errorMessage: { control: 'text' },
     handleSearchValue: { control: 'action' },
+    debounceDelay: { control: 'number' },
   },
 } satisfies Meta<typeof Search>
 
@@ -44,6 +45,7 @@ export const WithIcon: SearchStory = {
     searchIconName: 'search',
     placeholder: 'Search bar with icon',
     handleSearchValue: () => {},
+    debounceDelay: 500,
   },
 }
 
@@ -55,6 +57,7 @@ export const WithSvg: SearchStory = {
     searchSvgIcon: searchSvg,
     placeholder: 'Search bar with svg',
     handleSearchValue: () => {},
+    debounceDelay: 500,
   },
 }
 
@@ -67,6 +70,7 @@ export const WithLabel: SearchStory = {
     searchIconName: 'search',
     placeholder: 'Search bar with label',
     handleSearchValue: () => {},
+    debounceDelay: 500,
   },
 }
 
@@ -80,5 +84,6 @@ export const WithError: SearchStory = {
     isSearchError: true,
     errorMessage: 'Search error',
     handleSearchValue: () => {},
+    debounceDelay: 500,
   },
 }