+++ web/usuaris/src/__tests__/molecules/RoleFilter.test.tsx
@@ -3,8 +3,19 @@ import { fireEvent, render, screen, waitFor } from '../test-utils'
 import { RoleFilter } from '../../components/molecules/RoleFilter'
 import { UserRole } from '../../types'
 import { roles } from '../../constants'
+import { TAuthContext, useAuth } from '../../context/AuthProvider'
 
 const mockHandleClick = vi.fn()
+
+beforeEach(() => {
+  vi.mock('../../context/AuthProvider', async () => {
+    const actual = await vi.importActual('../../context/AuthProvider')
+    return {
+      ...actual,
+      useAuth: vi.fn(),
+    }
+  })
+})
 
 afterEach(() => {
   vi.restoreAllMocks()
@@ -12,12 +23,28 @@ afterEach(() => {
 
 describe('RolesFilter', () => {
   it('renders correctly', () => {
+    vi.mocked(useAuth).mockReturnValue({
+      user: {
+        dni: '12345678A',
+        email: 'user@example.cat',
+        role: 'ADMIN',
+      },
+    } as TAuthContext)
+
     render(<RoleFilter handleRole={mockHandleClick} />)
 
     waitFor(() => expect(screen.getByText(/rol/i)).toBeInTheDocument())
   })
 
   it('renders RoleList options when dropdown is clicked', async () => {
+    vi.mocked(useAuth).mockReturnValue({
+      user: {
+        dni: '12345678A',
+        email: 'user@example.cat',
+        role: 'ADMIN',
+      },
+    } as TAuthContext)
+
     render(<RoleFilter handleRole={mockHandleClick} />)
 
     const dropdown = screen.getByTestId('dropdown-header')
@@ -32,6 +59,14 @@ describe('RolesFilter', () => {
   })
 
   it('calls handleRole with the correct role when an option is clicked', () => {
+    vi.mocked(useAuth).mockReturnValue({
+      user: {
+        dni: '12345678A',
+        email: 'user@example.cat',
+        role: 'ADMIN',
+      },
+    } as TAuthContext)
+
     render(<RoleFilter handleRole={mockHandleClick} />)
 
     const dropdown = screen.getByTestId('dropdown-header')
@@ -47,5 +82,22 @@ describe('RolesFilter', () => {
       name: 'Administrador',
       slug: UserRole.ADMIN,
     })
+  })
+
+  it('disables roles filter when user mentor is logged', () => {
+    vi.mocked(useAuth).mockReturnValue({
+      user: {
+        dni: '12345678A',
+        email: 'user@example.cat',
+        role: 'MENTOR',
+      },
+    } as TAuthContext)
+
+    render(<RoleFilter handleRole={mockHandleClick} />)
+
+    const dropdownRole = screen.getByTestId('dropdown-header')
+    expect(dropdownRole).toBeInTheDocument()
+    expect(dropdownRole).toHaveTextContent(/registrat/i)
+    expect(dropdownRole).toBeDisabled()
   })
 })