+++ packages/ui/CHANGELOG.md
@@ -2,11 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.35.1] - 2024-06-19
+## [0.36.1] - 2024-06-19
 
 ### Fixed
 
 - Fixed corner backgroundImg of CardHome
+
+### Added
+
+## [0.36.0] - 2024-06-03
+
+- Added prop disabled to VoteCounter
+
+## [0.35.1] - 2024-06-18
+
+### Changed
+
+- Added debounceDelay as prop to Search
 
 ## [0.35.0] - 2024-06-10